requirements.txt
streamlit
pandas
openpyxl
python-dotenv
numpy
pillow
watchdog